Reading

HTMLTables.get — Function

HTMLTables.get(source::String; id::String="", classes::Union{Vector{String},String}="", index::Int=1)

Returns an HTML table a source as a string.

Arguments

source::String: URL or path to the HTML table.

id::String: The id of the HTML table.

classes::Union{Vector{String},String}: The classes of the HTML table.

index::Int: The index of the HTML table in the HTML document.

HTMLTables.getall — Function

HTMLTables.getall(source::String)

Extracts all tables from an HTML document or website.

Arguments

source::String: URL or path to the HTML document or website.

HTMLTables.read — Function

HTMLTables.read(source::String, sink::Function; id::String="", classes::Union{Vector{String},String}="")

Reads a HTML table into a sink function such as DataFrame.

Arguments

source::String: URL or path to the HTML table.

sink: The function that materializes the table data.

id::String: The id of the HTML table.

classes::Union{Vector{String},String}: The classes of the HTML table.

index::Int: The index of the HTML table in the HTML document.

Examples

# reading an HTML table into a DataFrame
using HTMLTables, DataFrames

df = HTMLTables.read("https://www.w3schools.com/html/html_tables.asp", DataFrame)

# writing the html table data into a CSV file
using CSV

CSV.write("table.csv", df)

# writing the html table data into a JSON file
using JSON, JSONTables

json = JSONTables.objecttable(df)

Base.open("table.json", "w") do io
    JSON3.pretty(io, json)
end

# writing the html table data into an Excel file
using XLSX

XLSX.writetable("table.xlsx", "Sheet 1" => df)

HTMLTables.readall — Function

HTMLTables.readall(source::String, sink::Function)

Reads all HTML tables into a sink function such as DataFrame.

Arguments

source::String: URL or path to the HTML document or website.

sink: The function that materializes the table data.